:recycle: Suggestions by Qodana (JetBrains)

+++ api/src/main/java/de/mymiggi/movie/api/actions/auditlog/GetAuditLogPageCountAction.java
@@ -1,15 +1,14 @@
 package de.mymiggi.movie.api.actions.auditlog;
 
-import jakarta.ws.rs.core.Response;
-
 import de.mymiggi.movie.api.entity.config.DefaultPage;
 import de.mymiggi.movie.api.entity.db.AuditLogEntity;
+import jakarta.ws.rs.core.Response;
 
 public class GetAuditLogPageCountAction
 {
 	public Response run(DefaultPage defaultPage)
 	{
-		double max = AuditLogEntity.count() / defaultPage.Size();
+		double max = ((double)AuditLogEntity.count()) / defaultPage.Size();
 		return Response.ok((int)Math.nextUp(max)).build();
 	}
 }
